Added NEW
Also added a TON of extra RAM to the input buffer; this is because what I'm going to be doing next involves some find-&-replace (i.e. PI for 3.14159, RC for the return code of system(), RND(min, max) for a random number etc.) and I don't want users to i.e. type PI a zillion times and crash the program.