Added a config file
Yes, I got tired of constantly having to start out by doing:

LOAD SCRIPTS/ALIASES
RUN
NEW

Every time I start a terminal.  Config is not a bad thing.  But it can be overdone.  If I find I need to go somewhere besides /etc, I will, but it's supposed to just be ONE config file.  Not one in /etc and one in ~ and 3 or 4 more in /etc/breakaway/conf or whatever the puck bigger programs do.  No.  ONE.  CONFIG.  FILE. :)