Fixed a minor memory error
I noticed sometimes when I would run system commands (like LS, using a "FILES" alias I have defined in my auto-run file) I get crashes with cryptic memory errors.  As much as I love C, its ability to report errors leaves much to be desired.  This is understandable, considering how low-level C really is (I still like calling it a "mid-level" language - higher than Assembly, lower than Python or JS).  The code is compiled to a binary format, so just the fact that the OS reports whether it was a malloc or a free that failed is... borderline helpful-ish? :D

Anyway, RunSYS sets the "RC" variable to be the result of the system program.  And in vars.c there is a place where I free var->name before freeing var.  That's really all this commit is.  Baby steps. :)