Starting over again (kinda)
The heap is the root of all evil... just kidding!  The love of money is the root of all evil... the heap just sucks ice! :D

I'm hanging on to all my old code, cuz I do plan to reuse some of it, but this time it's gonna be in ANSI C, super-portable and (hopefully) super-stable.  One thing I've noticed is that other projects - projects where I don't do a lot of malloc-&-free, are just plain better.  Less crashy.  Easier to debug, easier to write, and they just don't crash (and when they do, it's because I did something ridiculously stupid - I'll get to that later, lol).  Obviously, there are times when the heap is necessary, but as @Dirt2901 pointed out, having a fixed depth is an easy solution to memory-related problems.

Like previously, my little BASIC's program memory was solid - an array of strings.  Sure the array itself was on the heap, cuz I thought it would have to be, but editing it was super-easy, and the resulting code was pretty darn stable.

Now on to the stupid thing I did and didn't realize (lol): Oirinally I thought I was allocating 64000 bytes - but I was allocating 64000 * 80 bytes (cuz it's an array of LINES) which is why I was getting segfaults.  Oh, and I finally learned Valgrind! :)