Started on the manual.
I'm starting to... not lose interest int he project so much as get to the end of what I feel like spending a lot of time on.  Like most of the stuff on my GitHub, this is after-hours work, after all.  So I've added 7 things I may or may not add before calling version 1.0 done.  Breakaway BASIC is already my shell (replacing the awesome Gforth) and that's kinda what I wanted.

But... there sure are a lot of wouldn't-it-be-cool-if type things thare are not a monumental effort to make happen, so those are what I'm gonna shoot for next.